chore: integrate permissions policy with new model

+++ tests/users/test_api.py
@@ -4,8 +4,8 @@ import pytest
 from syrupy.matchers import path_type
 from virtool_core.models.enums import Permission
 
-from virtool.authorization.permissions import SpacePermission
-from virtool.authorization.relationships import UserPermission
+from virtool.authorization.relationships import SpaceUserRoleAssignment
+from virtool.authorization.roles import SpaceResourceRole
 from virtool.data.utils import get_data_from_app
 from virtool.groups.oas import UpdateGroupRequest, UpdatePermissionsRequest
 from virtool.settings.oas import UpdateSettingsRequest
@@ -238,13 +238,15 @@ class TestUpdate:
 
 @pytest.mark.parametrize("user", ["test", "bob"])
 async def test_list_permissions(spawn_client, user, snapshot):
-    client = await spawn_client(authorize=True, permissions=[Permission.create_sample, Permission.create_ref])
+    client = await spawn_client(
+        authorize=True, permissions=[Permission.create_sample, Permission.create_ref]
+    )
 
     authorization_client = client.app["authorization"]
 
     await authorization_client.add(
-        UserPermission("test", SpacePermission.CREATE_SAMPLE),
-        UserPermission("test", SpacePermission.CREATE_REFERENCE),
+        SpaceUserRoleAssignment(0, "test", SpaceResourceRole.SAMPLE_EDITOR),
+        SpaceUserRoleAssignment(0, "test", SpaceResourceRole.REFERENCE_BUILDER),
     )
 
     resp = await client.get(
@@ -256,44 +258,34 @@ async def test_list_permissions(spawn_client, user, snapshot):
 
 
 @pytest.mark.parametrize(
-    "permission, status",
+    "role, status",
     [
-        (SpacePermission.CREATE_SAMPLE.value.id, 200),
+        (SpaceResourceRole.SAMPLE_EDITOR, 200),
         ("invalid", 400),
     ],
-    ids=[
-        "valid_permission",
-        "invalid_permission"
-    ],
+    ids=["valid_permission", "invalid_permission"],
 )
-async def test_add_permission(spawn_client, permission, status, snapshot):
+async def test_add_permission(spawn_client, role, status, snapshot):
     client = await spawn_client(authorize=True, administrator=True)
 
-    resp = await client.put(
-        f"/users/test/permissions/{permission}", {}
-    )
+    resp = await client.put(f"/users/test/permissions/{role}", {})
 
     assert resp.status == status
     assert await resp.json() == snapshot()
 
 
 @pytest.mark.parametrize(
-    "permission, status",
+    "role, status",
     [
-        (SpacePermission.CREATE_SAMPLE.name, 200),
+        (SpaceResourceRole.SAMPLE_EDITOR, 200),
         ("invalid", 400),
     ],
-    ids=[
-        "valid_permission",
-        "invalid_permission"
-    ],
+    ids=["valid_permission", "invalid_permission"],
 )
-async def test_remove_permission(spawn_client, permission, status, snapshot):
+async def test_remove_permission(spawn_client, role, status, snapshot):
     client = await spawn_client(authorize=True, administrator=True)
 
-    resp = await client.delete(
-        f"/users/test/permissions/{permission}"
-    )
+    resp = await client.delete(f"/users/test/permissions/{role}")
 
     assert resp.status == status
     assert await resp.json() == snapshot()
